requirements.txt
pandas
scipy
scikit-learn
clickhouse-driver
python-dotenv
numpy
plotly
flask
openvpn-api
waitress
bs4
requests